Add script to get metadata from directory names
Returns a set of arguments suitable for passing to AtomicParsley when
given a directory name in a specific format. Currently only accepts
TV shows in the format:

    Show Name - 4x22 - Episode Title

Any show which is not an episode (eg 22) within a season (eg 4) and/or
without a title (ie specials, one-offs) won't be understood.