Adding multiple DVDs should add their tracks
When multiple DVDs were passed as arguments to `media-add`, the global
`tracks` variable was mistakenly being over written by not being a local
variable in `convert_and_add_from_dvd_image`. This caused the second DVD
to try to use the same tracks as the first, which caused problems when
any of those tracks didn't appear (it also meant tracks that should be
converted were not). This fixes the bug.