Fix testing of overriding DVD HandBrake settings
The `testing_handbrake_args` function didn't respect the metadata
file settings, so whilst DVDs could be overridden it didn't show up
in testing.

This simplifies the behaviour when setting `MEDIA_TESTING` to be the
lowering of quality and only encoding one minute of content. Nothing
was actually using the `small` testing settings anyway.