Add a script to preprocess FLAC format albums
Converts the FLAC files in a directory to WAV files, and creates a
metadata file, ready for add-cd to process. Where the filenames are in
the right format a mostly-meaningful metadata file will be populated.

Relocates the `apply_title_case` function from create-audio-metadata
to the main media script so it can be reused.