Add video_lang setting to configuration tests
This was skipped when adding the setting, but the tests continued
to pass because it was the last line of the output of `show-config`
and the test only tested the lines that matched the expected output.

Now includes a test that the number of lines of output and expected
output match, which would have caused the test to fail once the
video_lang setting was added to the code but not the test. This should
catch future new configuration being added without updating this test.