Add code to use CDDB inexact matches
Some CD lookups give a return code of 211 -- inexact match. In my
sample size of one example so far (Falco's Einzelhaft), it was the
right result, and only the one result, so we treat it as a 200
exact match.

Required a small fix where `rip-cd` was was using the ID returned by
CDDB as the destination directory (and not the calculated ID), which
broke other assumptions.