Add album and artist metadata to TV shows
There is an intermittent bug in iTunes' handling of user created
TV shows where updating the watched status of one episode will
affect the metadata of another with the same title. For example,
adding a new show's "Pilot" episode will have it marked as watched
when added, and manually marking it as unwatched will trigger
every other episode with the title "Pilot" (which is probably a
lot of episodes) to appear unwatched too.

TV downloaded from iTunes has both the album and artist metadata set
to the combo "SERIES, Season SEASON". Yet when examining the metadata
via "Show Programme Info", neither album nor artist is available as
metadata to change, implying that it is not necessary.

However, as a working hypothesis, the lack of album (or artist?)
metadata is what is causing iTunes to be confused as to which "Pilot"
episode is which. Therefore let's add it to all newly minted shows
until such time as the hypothesis is proven wrong.